requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.7
fastavro==1.4.5
idna==3.3
pulsar-client==2.8.1
python-consul==1.1.0
requests==2.26.0
six==1.16.0
urllib3==1.26.7
wakeonlan==2.0.1
